Fix PostgreSQL schema dialect bugs + add dual-engine CI matrix
- 010: parenthesize nzbdav_id expression index (was 42601, aborted migration chain)
- 001/016: INTEGER -> BIGINT for byte/size columns
- 027: file_health.metadata JSONB -> TEXT
- 032: idempotent heal migration for pre-existing Postgres DBs
- CI: run dialect-parity suite against SQLite and Postgres

+++ internal/database/migrations/postgres/016_add_import_hourly_stats.sql
@@ -3,7 +3,7 @@ CREATE TABLE IF NOT EXISTS import_hourly_stats (
     hour TIMESTAMPTZ PRIMARY KEY,
     completed_count INTEGER DEFAULT 0,
     failed_count INTEGER DEFAULT 0,
-    bytes_downloaded INTEGER DEFAULT 0,
+    bytes_downloaded BIGINT DEFAULT 0,
     updated_at TIMESTAMPTZ DEFAULT CURRENT_TIMESTAMP
 );
